Make nightly dashboard trend plot interactive (#64)
This PR is a draft to address #55. To do this I have ported the report
to become a quarto doc rather than rmarkdown and then written the viz in
javascript for more interactivity.

Because the way this is implemented, it needs to be served up via https
rather than a local html file. So screenshots it is. Here is the default
which sets the x-axes to extend to the last 120 days but we can slide to only look at the last ten days or look at the past 6 months.

I have also updated the build table to include passing runs. Because
this adds a significant amount of rows to the table, I've implemented
some interactivity for the build table. This looks like this: