Visualize and interact with micro benchmarks in performance report (#65)
This PR implements an interface to work with the microbenchmarks,
include the C++ and Java ones. Additionally, because there are _so_
many, I've enabled the user to be able to filter for conbench detected
regressions and improvements. I've followed the style of the other plots
in this report though these one are built in Javascript.

I have also added a link so users can go to conbench to directly view
the compare page.

---------

Co-authored-by: Edward Visel <1693477+alistaire47@users.noreply.github.com>