Add JS Benchmarks to Performance Report (#67)
After #65 I realized that it would be simple to add JavaScript
benchmarks (which we've not visualized before). Because those are run in
the same machines as the R and python benchmarks I've added a column
that distinguishes between the macro/micro benchmarks:


![image](https://github.com/ursacomputing/crossbow/assets/18472598/85253c94-9117-4097-8bf6-8371466dd906)

Otherwise this PR is just some display tweaks to better show labels and
updating some text to better explain the microbenchmark explorer.